// GYMFRONTEND/src/app/core/interfaces/classes/classes.ts
export interface Classes {

       id:string;
       classname: string;
       classday: string; // Monday- tuesday .....
       dateBegin: Date; // date when the class begins
       code: string ; // Class code
       classtime: string; // 10:00 am - 11:00 am 
       classlevel: string; // low - medium - high
       session_time: number; // days
       price: number; // price per month
       trainer: string;
       key_benefits: string;
       expert_trainer: string;
       class_overview: string;
       why_matters: string;
       image: string;

}